EasySoft/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai>=1.40.0
weaviate-client==4.5.2
python-dotenv==1.0.0
numpy==1.24.3
waitress==2.1.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Azure dependencies
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-storage-file-share==12.14.0